pyproject.toml
[tool.poetry]
name = "cac-core"
version = "0.1.0"
description = ""
authors = ["Ryan Punt <ryan@mirum.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
logger = "^1.4"
tabulate = "^0.9.0"
requests = "^2.32.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
